move encrypt_op{_sev} into system ioctls
Based on the KVM API, KVM_MEMORY_ENCRYPT_OP is a system ioctl. However,
this was not reflected in the crate.
